smart-door-system/requirements.txt
boto3
firebase-admin
opencv-python
paho-mqtt
RPi.GPIO
